[Fix] Signup MVVM
