+++ mohaemukzip/Sources/Feature/Onboarding/Models/SignupModel.swift
@@ -1,0 +1,14 @@
+//
+//  SignupModel.swift
+//  mohaemukzip
+//
+//  Created by 이서현 on 2/2/26.
+//
+
+
+struct SignupModel: Equatable {
+    var nickname: String = ""
+    var userId: String = ""
+    var password: String = ""
+    var passwordConfirm: String = ""
+}
