+++ mohaemukzip/Sources/Feature/Onboarding/Views/SignupView.swift
@@ -8,150 +8,6 @@
 import SwiftUI
 import Observation
 
-// MARK: - Model
-struct SignupModel: Equatable {
-    var nickname: String = ""
-    var userId: String = ""
-    var password: String = ""
-    var passwordConfirm: String = ""
-}
-
-// MARK: - ViewModel
-@Observable
-final class SignupViewModel {
-
-    enum IdCheckState: Equatable {
-        case none
-        case invalidFormat
-        case duplicated
-        case available
-
-        var message: String? {
-            switch self {
-            case .none: return nil
-            case .invalidFormat: return "아이디 형식이 올바르지 않아요."
-            case .duplicated: return "이미 사용 중인 아이디예요. 다른 아이디를 입력해 주세요."
-            case .available: return "사용 가능한 아이디예요."
-            }
-        }
-
-        var isError: Bool {
-            switch self {
-            case .invalidFormat, .duplicated:
-                return true
-            default:
-                return false
-            }
-        }
-
-        var isSuccess: Bool {
-            self == .available
-        }
-    }
-
-    var model: SignupModel = .init()
-
-    // UI State
-    var idCheckState: IdCheckState = .none
-    var isPasswordMismatch: Bool = false
-
-    func onChangeNickname(_ newValue: String) {
-        // 15자 제한
-        if newValue.count <= 15 {
-            model.nickname = newValue
-        } else {
-            model.nickname = String(newValue.prefix(15))
-        }
-    }
-
-    func onChangeUserId(_ newValue: String) {
-        // 15자 제한 (15자 넘어가면 더 이상 타이핑 안 되게)
-        if newValue.count <= 15 {
-            model.userId = newValue
-        } else {
-            model.userId = String(newValue.prefix(15))
-        }
-        // 아이디가 바뀌면 중복확인 다시 필요
-        idCheckState = .none
-    }
-
-    func onChangePassword(_ newValue: String) {
-        model.password = newValue
-        validatePasswordMatch()
-    }
-
-    func onChangePasswordConfirm(_ newValue: String) {
-        model.passwordConfirm = newValue
-        validatePasswordMatch()
-    }
-
-    func checkDuplicateId() {
-        guard isValidIdFormat(model.userId) else {
-            idCheckState = .invalidFormat
-            return
-        }
-
-        // 예: "abcde"는 사용 가능, "taken"은 중복, 나머지는 사용 가능 처리
-        let lower = model.userId.lowercased()
-        if lower == "taken" || lower == "admin" {
-            idCheckState = .duplicated
-        } else {
-            idCheckState = .available
-        }
-    }
-
-    var nicknameCountText: String {
-        "\(model.nickname.count)/15"
-    }
-
-    var userIdCountText: String {
-        "\(model.userId.count)/15"
-    }
-
-    var isValidNickname: Bool {
-        !model.nickname.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-    }
-
-    var isValidPassword: Bool {
-        isValidPasswordFormat(model.password)
-    }
-
-    var isReadyToStart: Bool {
-        isValidNickname
-        && idCheckState == .available
-        && isValidPassword
-        && !isPasswordMismatch
-        && !model.passwordConfirm.isEmpty
-    }
-
-    func helperTextForIdFormat() -> String {
-        "6~15자 영문, 숫자를 입력해주세요."
-    }
-
-    func helperTextForPasswordFormat() -> String {
-        "6~20자 영문, 숫자, 특수문자를 입력해주세요."
-    }
-
-    private func validatePasswordMatch() {
-        guard !model.password.isEmpty || !model.passwordConfirm.isEmpty else {
-            isPasswordMismatch = false
-            return
-        }
-        isPasswordMismatch = (!model.passwordConfirm.isEmpty) && (model.password != model.passwordConfirm)
-    }
-
-    private func isValidIdFormat(_ text: String) -> Bool {
-        // 6~15자 영문/숫자
-        let pattern = "^[A-Za-z0-9]{6,15}$"
-        return text.range(of: pattern, options: .regularExpression) != nil
-    }
-
-    private func isValidPasswordFormat(_ text: String) -> Bool {
-        let pattern = "^(?=.*[A-Za-z])(?=.*[0-9])(?=.*[^A-Za-z0-9]).{6,20}$"
-        return text.range(of: pattern, options: .regularExpression) != nil
-    }
-}
-
 struct SignupView: View {
 
     @EnvironmentObject private var router: AuthRouter
@@ -266,9 +122,9 @@ struct SignupView: View {
                 }
 
                 Button {
-                    viewModel.checkDuplicateId()
+                    Task { await viewModel.checkDuplicateId() }
                 } label: {
-                    Text("중복확인")
+                    Text(viewModel.isCheckingId ? "확인중" : "중복확인")
                         .font(.PretendardMedium16)
                         .foregroundStyle(.main400)
                         .frame(width: 84, height: 44)
@@ -278,9 +134,11 @@ struct SignupView: View {
                         )
                         .contentShape(Rectangle())
                 }
+                .disabled(viewModel.isCheckingId)
+                .opacity(viewModel.isCheckingId ? 0.6 : 1)
             }
 
-            if let msg = viewModel.idCheckState.message {
+            if let msg = (viewModel.idCheckMessage ?? viewModel.idCheckState.message) {
                 HStack(spacing: 6) {
                     Image(systemName: viewModel.idCheckState.isSuccess ? "checkmark.circle.fill" : "exclamationmark.circle.fill")
                         .font(.system(size: 12, weight: .semibold))
@@ -361,7 +219,18 @@ struct SignupView: View {
                 .font(.PretendardRegular16)
             }
 
-            if viewModel.isPasswordMismatch {
+            if let error = viewModel.errorMessage {
+                HStack(spacing: 6) {
+                    Image(systemName: "exclamationmark.circle.fill")
+                        .font(.PretendardRegular13)
+                        .foregroundStyle(Color.red)
+
+                    Text(error)
+                        .font(.PretendardRegular13)
+                        .foregroundStyle(Color.red)
+                }
+                .padding(.top, 2)
+            } else if viewModel.isPasswordMismatch {
                 HStack(spacing: 6) {
                     Image(systemName: "exclamationmark.circle.fill")
                         .font(.PretendardRegular13)
@@ -378,17 +247,22 @@ struct SignupView: View {
 
     private var startButton: some View {
         Button {
-            router.push(.signupFinish)
+            Task {
+                let ok = await viewModel.signup()
+                if ok {
+                    router.push(.signupFinish)
+                }
+            }
         } label: {
             Text("시작하기")
                 .font(.PretendardSemibold18 )
-                .foregroundStyle(viewModel.isReadyToStart ? Color.white : Color.white.opacity(0.9))
+                .foregroundStyle(viewModel.isReadyToStart && !viewModel.isLoading ? Color.white : Color.white.opacity(0.9))
                 .frame(maxWidth: .infinity)
                 .frame(height: 57)
-                .background(viewModel.isReadyToStart ? .main400 : .grey400)
+                .background((viewModel.isReadyToStart && !viewModel.isLoading) ? .main400 : .grey400)
                 .clipShape(RoundedRectangle(cornerRadius: 10))
                 .contentShape(Rectangle())
         }
-        .disabled(!viewModel.isReadyToStart)
+        .disabled(!viewModel.isReadyToStart || viewModel.isLoading)
     }
 }
